Change summons entity super class to BasicEntitySummon. (except mounts) Rewrite all attackWithAmmo and attackedFrom method. Add some damage source checking.

+++ src/main/java/com/lulan/shincolle/item/ShipSpawnEgg.java
@@ -453,7 +453,7 @@ public class ShipSpawnEgg extends BasicItem
                         
                     	if (ship != null)
                     	{
-                    		ship.initAttrs((byte) player.getRNG().nextInt(4));
+                    		ship.initAttrs(player.getRNG().nextInt(4));
                         	player.world.spawnEntity(ship);
                         	ship.playLivingSound();
                     	}
